% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tsv_data.R
\name{infer_merged_bam_names}
\alias{infer_merged_bam_names}
\title{Infer the bam names for different merged states from a TSV parsed with \link[eagerR]{read_input_tsv_data}}
\usage{
infer_merged_bam_names(parsed_tsv, run_trim_bam = NA)
}
\arguments{
\item{parsed_tsv}{Tibble. A tibble generated using \link[eagerR]{read_input_tsv_data}.}

\item{run_trim_bam}{logical. Was `--run_trim_bam` used in the eager run?}
}
\value{
A Tibble containing two additional columns with the names of the merged bams for a sample after each merging step.
}
\description{
Infer the bam names for different merged states from a TSV parsed with \link[eagerR]{read_input_tsv_data}
}
