% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eager_result_files.R
\name{read_angsd_cont_json}
\alias{read_angsd_cont_json}
\title{Read nf-core/eager nuclear contamination results JSON}
\usage{
read_angsd_cont_json(angsd_cont_json)
}
\arguments{
\item{angsd_cont_json}{character. The path to the input nuclear contamination result json}
}
\value{
A Tibble containing the per-library nuclear contamination results.
}
\description{
Read nf-core/eager nuclear contamination results JSON
}
