% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eager_result_files.R
\name{read_damageprofiler_json}
\alias{read_damageprofiler_json}
\title{Read nf-core/eager DamageProfiler results JSON}
\usage{
read_damageprofiler_json(dmgprof_json, library_id = "")
}
\arguments{
\item{dmgprof_json}{character. The path to the input DamageProfiler result json for the specified library.}

\item{library_id}{character. The Library_ID of the library the results are based on.}
}
\value{
A Tibble containing the DamageProfiler results for the specified library.
}
\description{
Read nf-core/eager DamageProfiler results JSON
}
