% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eager_result_files.R
\name{read_damageprofiler_jsons_from_dir}
\alias{read_damageprofiler_jsons_from_dir}
\title{Read all Damageprofiles JSONs in a DamageProfiler results directory into a tibble.}
\usage{
read_damageprofiler_jsons_from_dir(result_dir)
}
\arguments{
\item{result_dir}{character. The path to the root DamageProfiler results directory for an eager run.}
}
\value{
A Tibble containing the JSON results (compiled using \link[eagerR]{read_damageprofiler_json}) for all
    libraries in the eager run.
}
\description{
Read all Damageprofiles JSONs in a DamageProfiler results directory into a tibble.
}
