% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eager_result_files.R
\name{read_endorspy_json}
\alias{read_endorspy_json}
\title{Read nf-core/eager endorspy results JSON}
\usage{
read_endorspy_json(endorspy_json, library_id = "")
}
\arguments{
\item{endorspy_json}{character. The path to the input endorspy result json for the specified library.}

\item{library_id}{character. The Library_ID of the library the results are based on.}
}
\value{
A Tibble containing the endorspy results for the specified library.
}
\description{
Read nf-core/eager endorspy results JSON
}
