% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eager_result_files.R
\name{read_endorspy_jsons_from_dir}
\alias{read_endorspy_jsons_from_dir}
\title{Read all endorspy JSONs in an endorspy results directory into a tibble.}
\usage{
read_endorspy_jsons_from_dir(result_dir)
}
\arguments{
\item{result_dir}{character. The path to the endorspy results directory for an eager run.}
}
\value{
A Tibble containing the JSON results (compiled using \link[eagerR]{read_endorspy_json}) for all
    libraries in the eager run.
}
\description{
Read all endorspy JSONs in an endorspy results directory into a tibble.
}
