% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tsv_data.R
\name{read_input_tsv_data}
\alias{read_input_tsv_data}
\title{Read the data in an eager input TSV file}
\usage{
read_input_tsv_data(tsv_fn)
}
\arguments{
\item{tsv_fn}{character. The path to the input TSV file used in your eager run}
}
\value{
A tibble containing the eager TSV information. This tibble includes three additional columns: initial_merge additional_merge strandedness_clash.
   The initial_merge column is logical specifying if merging of libraries of the same UDG/strandedness combination was done after deduplication.
   The additional_merge column is logical specifying if merging of libraries of the same strandedness but different UDG was done after bam trimming.
   The strandedness_clash column is a logical specifying if libraries of differing strandedness exist that share the same Sample_Name.
}
\description{
Read the data in an eager input TSV file
}
