% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eager_result_files.R
\name{read_sexdet_json}
\alias{read_sexdet_json}
\title{Read nf-core/eager Sex Determination results JSON}
\usage{
read_sexdet_json(sexdet_json)
}
\arguments{
\item{sexdet_json}{character. The path to the input sex determination result json}
}
\value{
A Tibble containing the sex determination results.
}
\description{
Read nf-core/eager Sex Determination results JSON
}
