% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eager_result_files.R
\name{read_snp_coverage_json}
\alias{read_snp_coverage_json}
\title{Read nf-core/eager EIGENSTRAT SNP coverage results JSON}
\usage{
read_snp_coverage_json(snp_cov_json)
}
\arguments{
\item{snp_cov_json}{character. The path to the input eigenstrat snp coverage result json}
}
\value{
A Tibble containing the per-library eigenstrat snp coverage results.
}
\description{
Read nf-core/eager EIGENSTRAT SNP coverage results JSON
}
